extracted reusable stats-parts into a trait
